chore(WebGPU): move two more tests into Rendering\Core
move sphere and stick mapper tests into Core
