+++ Sources/Rendering/Core/SphereMapper/test/testSphere.png
