+++ Sources/Rendering/Core/StickMapper/test/testStick.png
