+++ Sources/tests.js
@@ -53,6 +53,8 @@ import './Rendering/Core/PointPicker/test/testPointPicker';
 import './Rendering/Core/Prop3D/test/testUserMatrix';
 import './Rendering/Core/RenderWindow/test/testMultipleRenderers';
 import './Rendering/Core/SphereMapper/test/testDisableScalarColoring';
+import './Rendering/Core/SphereMapper/test/testSphere';
+import './Rendering/Core/StickMapper/test/testStick';
 import './Rendering/Misc/GenericRenderWindow/test/testGenericRenderWindowCreateDelete';
 import './Rendering/OpenGL/HardwareSelector/test/testHardwareSelector';
 import './Rendering/OpenGL/HardwareSelector/test/testHardwareSelectorGlyph';
@@ -64,10 +66,8 @@ import './Rendering/OpenGL/PolyDataMapper/test/testAddShaderReplacements';
 import './Rendering/OpenGL/PolyDataMapper/test/testClearShaderReplacements';
 import './Rendering/OpenGL/PolyDataMapper/test/testClippingPlanes';
 import './Rendering/OpenGL/PolyDataMapper/test/testInterpolateScalarsBeforeMapping';
-import './Rendering/OpenGL/SphereMapper/test/testSphere';
 import './Rendering/OpenGL/Skybox/test/testSkybox';
 import './Rendering/OpenGL/Skybox/test/testSkyboxBackground';
-import './Rendering/OpenGL/StickMapper/test/testStick';
 import './Rendering/OpenGL/Texture/test/testCreateCubeFromRawTexture';
 import './Rendering/OpenGL/VolumeMapper/test/testComposite';
 import './Rendering/OpenGL/VolumeMapper/test/testCompositeParallelProjection';
